requirements.txt
# ===== CORE DEPENDENCIES =====
streamlit==1.39.0
pandas==2.1.4
numpy==1.26.4

# ===== GOOGLE SHEETS API =====
gspread==6.1.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# ===== PDF GENERATION =====
reportlab==4.2.5
Pillow==10.4.0

# ===== UTILITIES =====
python-dateutil==2.9.0
requests==2.32.3

# ===== OPTIONAL (si WhatsApp está activo) =====
# python-dotenv==1.0.1